+++ src/net/server.rs
@@ -9,6 +9,8 @@ use {
         fs,
         io::BufReader,
         os::unix::net::UnixListener,
+        path::PathBuf,
+        sync::{Arc, Mutex},
         thread,
     },
 };
@@ -18,7 +20,11 @@ pub struct Server {
 }
 
 impl Server {
-    pub fn new(name: &str, tx: Sender<Sequence>) -> Result<Self, NetError> {
+    pub fn new(
+        name: &str,
+        tx: Sender<Sequence>,
+        root: Arc<Mutex<PathBuf>>,
+    ) -> Result<Self, NetError> {
         let path = super::socket_file_path(name);
         if fs::metadata(&path).is_ok() {
             return Err(NetError::SocketNotAvailable { path });
@@ -30,13 +36,23 @@ impl Server {
         thread::spawn(move || {
             for stream in listener.incoming() {
                 match stream {
-                    Ok(stream) => {
-                        let mut br = BufReader::new(stream);
+                    Ok(mut stream) => {
+                        let mut br = BufReader::new(&stream);
                         if let Some(sequence) = match Message::read(&mut br) {
                             Ok(Message::Command(command)) => {
                                 debug!("got single command {:?}", &command);
                                 // we convert it to a sequence
                                 Some(Sequence::new_single(command))
+                            }
+                            Ok(Message::GetRoot) => {
+                                debug!("got get root query");
+                                let root = root.lock().unwrap();
+                                let answer = Message::Root(root.to_string_lossy().to_string());
+                                match answer.write(&mut stream) {
+                                    Ok(()) => debug!("root path successfuly returned"),
+                                    Err(e) => warn!("error while answering: {:?}", e),
+                                }
+                                None
                             }
                             Ok(Message::Sequence(sequence)) => {
                                 debug!("got sequence {:?}", &sequence);